+++ lib/src/variant/variant.dart
@@ -39,7 +39,7 @@ class Variant {
   /// Material conditions that define how insufficient material draws should be decided.
   final MaterialConditions<String> materialConditions;
 
-  final GameEndConditions gameEndConditions;
+  final GameEndConditionsSet gameEndConditions;
 
   final OutputOptions outputOptions;
 
@@ -113,7 +113,7 @@ class Variant {
     this.castlingOptions = CastlingOptions.standard,
     this.promotionOptions = PromotionOptions.standard,
     this.materialConditions = MaterialConditions.none,
-    this.gameEndConditions = GameEndConditions.standard,
+    this.gameEndConditions = GameEndConditionsSet.standard,
     this.outputOptions = OutputOptions.standard,
     this.startPosition,
     this.startPosBuilder,
@@ -140,7 +140,7 @@ class Variant {
     CastlingOptions? castlingOptions,
     PromotionOptions? promotionOptions,
     MaterialConditions<String>? materialConditions,
-    GameEndConditions? gameEndConditions,
+    GameEndConditionsSet? gameEndConditions,
     OutputOptions? outputOptions,
     String? startPosition,
     FenBuilder? startPosBuilder,
@@ -193,6 +193,7 @@ class Variant {
       castlingOptions: CastlingOptions.standard,
       materialConditions: MaterialConditions.standard,
       outputOptions: OutputOptions.standard,
+      promotionOptions: PromotionOptions.standard,
       enPassant: true,
       halfMoveDraw: 100,
       repetitionDraw: 3,
@@ -338,7 +339,7 @@ class Variant {
 
   factory Variant.threeCheck() => Variant.standard().copyWith(
         name: 'Three Check',
-        gameEndConditions: GameEndConditions.threeCheck,
+        gameEndConditions: GameEndConditionsSet.threeCheck,
       );
 
   factory Variant.kingOfTheHill() {
@@ -374,6 +375,13 @@ class Variant {
       ],
     );
   }
+
+  factory Variant.horde() => Variant.standard().copyWith(
+        name: 'Horde Chess',
+        startPosition:
+            'rnbqkbnr/pppppppp/8/1PP2PP1/PPPPPPPP/PPPPPPPP/PPPPPPPP/PPPPPPPP w kq - 0 1',
+        gameEndConditions: GameEndConditionsSet.horde,
+      );
 
   factory Variant.spawn() {
     final standard = Variant.standard();